Changed the Css upload functionality (#823)
