+++ app/Http/Resources/V1/ActivityTypeResource.php
@@ -20,7 +20,6 @@ class ActivityTypeResource extends JsonResource
             'title' => $this->title,
             'order' => $this->order,
             'image' => $this->image,
-            'css_path' => $this->css_path,
             'activityItems' => $this->activityItems,
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
